Fixed minor building warning in Windows
git-svn-id: https://svn.code.sf.net/p/veroroute/code/trunk@82 d5f823ff-b8c7-416f-ab5d-d69718ecf442

+++ src/GWriter.cpp
@@ -281,7 +281,7 @@ void GStream::OutLine(const Curve& curve, bool bForceClose)	// Outline of a curv
 {
 	if ( curve.empty() ) return;
 	SetPen(curve.m_pen);
-	const int N = curve.size();
+	const size_t N = curve.size();
 	if	( N == 1 ) return Flash( curve.front() );
 	if	( N == 2 ) return Line( curve.front(), curve.back() );
 	auto& front = curve.front();
